Do not use hardcoded Python and Node versions for CI (#184)
Python version is defined in pyproject.toml, Node version in package.json and their lock files.
With this proposed change, the versions are taken from these files instead of hardcoding them in the workflows.

The setup-env action gets the values from the files, so the inputs are no longer needed. The downside would be that we can't test against multiple versions (eg. via a matrix), but this was never a requirement and wasn't needed since the beginning of this project

Also enabled caching for node.